+++ app/src/main/java/com/dilekbaykara/tasky/di/AuthModule.kt
@@ -2,7 +2,7 @@ package com.dilekbaykara.tasky.di
 
 import android.util.Log
 import com.dilekbaykara.tasky.BuildConfig
-import com.dilekbaykara.tasky.domain.repository.AuthRepositoryImpl
+import com.dilekbaykara.tasky.data.repository.AuthRepositoryImpl
 import com.dilekbaykara.tasky.data.remote.TaskyApi
 import dagger.Module
 import dagger.Provides
